Move TestUtil to util package
