+++ app/src/androidTest/java/com/twilio/video/app/e2eTest/AppLinkRoomTest.kt
@@ -6,9 +6,9 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
 import androidx.test.rule.ActivityTestRule
 import com.twilio.video.app.R
-import com.twilio.video.app.getString
-import com.twilio.video.app.retrieveEmailCredentials
-import com.twilio.video.app.retryEspressoAction
+import com.twilio.video.app.util.getString
+import com.twilio.video.app.util.retrieveEmailCredentials
+import com.twilio.video.app.util.retryEspressoAction
 import com.twilio.video.app.screen.assertRoomNameIsDisplayed
 import com.twilio.video.app.screen.clickSettingsMenuItem
 import com.twilio.video.app.screen.loginWithEmail
