+++ app/src/androidTest/java/com/twilio/video/app/screen/SettingsScreen.kt
@@ -9,8 +9,8 @@ import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import com.twilio.video.app.R
-import com.twilio.video.app.getString
-import com.twilio.video.app.retryEspressoAction
+import com.twilio.video.app.util.getString
+import com.twilio.video.app.util.retryEspressoAction
 
 fun assertSettingsTitleIsVisible() {
     retryEspressoAction { onView(withText(getString(R.string.settings))).check(matches(isDisplayed())) }
